familjekontor/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Login==0.6.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
openpyxl==3.1.5
python-dotenv==1.0.1
Werkzeug==3.1.3
email-validator==2.2.0
WTForms==3.2.1
pytest==8.3.4
pytest-flask==1.3.0
